+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number comes later -->
+- feat: Deprecate `getModifierClasses` and `getBemClass` — in favor of
+  `{ modifiedClass } from '@hugsmidjan/qj/classUtils';`
 - feat: Optimize performance of `useClientState`/`useIsBrowserSide`/etc.
+- fix: Correctly remove `Portal`'s root element on unmount (affects `Modal`)
 
 ## 0.4.32